Add casts and remove unused variables to eliminate compiler warnings.
git-svn-id: http://svn.eol.ucar.edu/svn/eldora/trunk@1874 15f24ed5-0439-0410-88a4-bd1d0fc06cdd

+++ ecb/dectohex.cpp
@@ -17,7 +17,7 @@ char *dectohex(int dec)
 		dec/=16;
 		temp1 = temp1-dec;
 		temp1 *= 16;
-		temp2 = temp1;
+		temp2 = (int)temp1;
 		switch(temp2)
 			{
 			case 15:
